* Generic-crypto calls moved to dedicated crypto module (in progress)
