+++ libindy/src/api/mod.rs
@@ -2,6 +2,7 @@ extern crate libc;
 
 pub mod agent;
 pub mod anoncreds;
+pub mod crypto;
 pub mod signus;
 pub mod ledger;
 pub mod pool;
