+++ libindy/src/api/signus.rs
@@ -186,139 +186,6 @@ pub  extern fn indy_store_their_did(command_handle: i32,
     result_to_err_code!(result)
 }
 
-/// Creates keys pair and stores in the wallet.
-///
-/// #Params
-/// command_handle: Command handle to map callback to caller context.
-/// wallet_handle: Wallet handle (created by open_wallet).
-/// key_json: Key information as json. Example:
-/// {
-///     "seed": string, // Optional (if not set random one will be used); Seed information that allows deterministic key creation.
-///     "crypto_type": string, // Optional (if not set then ed25519 curve is used); Currently only 'ed25519' value is supported for this field.
-/// }
-/// cb: Callback that takes command result as parameter.
-///
-/// #Returns
-/// Error Code
-/// cb:
-/// - xcommand_handle: command handle to map callback to caller context.
-/// - err: Error code.
-/// - verkey: Ver key of generated key pair, also used as key identifier
-///
-/// #Errors
-/// Common*
-/// Wallet*
-/// Crypto*
-#[no_mangle]
-pub  extern fn indy_create_key(command_handle: i32,
-                               wallet_handle: i32,
-                               key_json: *const c_char,
-                               cb: Option<extern fn(xcommand_handle: i32, err: ErrorCode,
-                                                    verkey: *const c_char)>) -> ErrorCode {
-    check_useful_c_str!(key_json, ErrorCode::CommonInvalidParam3);
-    check_useful_c_callback!(cb, ErrorCode::CommonInvalidParam4);
-
-    let result = CommandExecutor::instance()
-        .send(Command::Signus(SignusCommand::CreateKey(
-            wallet_handle,
-            key_json,
-            Box::new(move |result| {
-                let (err, verkey) = result_to_err_code_1!(result, String::new());
-                let verkey = CStringUtils::string_to_cstring(verkey);
-                cb(command_handle, err, verkey.as_ptr())
-            })
-        )));
-
-    result_to_err_code!(result)
-}
-
-/// Saves/replaces the meta information for the giving key in the wallet.
-///
-/// #Params
-/// command_handle: Command handle to map callback to caller context.
-/// wallet_handle: Wallet handle (created by open_wallet).
-/// verkey - the key (verkey, key id) to store metadata.
-/// metadata - the meta information that will be store with the key.
-/// cb: Callback that takes command result as parameter.
-///
-/// #Returns
-/// Error Code
-/// cb:
-/// - xcommand_handle: command handle to map callback to caller context.
-/// - err: Error code.
-///
-/// #Errors
-/// Common*
-/// Wallet*
-/// Crypto*
-#[no_mangle]
-pub  extern fn indy_set_key_metadata(command_handle: i32,
-                                     wallet_handle: i32,
-                                     verkey: *const c_char,
-                                     metadata: *const c_char,
-                                     cb: Option<extern fn(xcommand_handle: i32,
-                                                          err: ErrorCode)>) -> ErrorCode {
-    check_useful_c_str!(verkey, ErrorCode::CommonInvalidParam3);
-    check_useful_c_str_empty_accepted!(metadata, ErrorCode::CommonInvalidParam4);
-    check_useful_c_callback!(cb, ErrorCode::CommonInvalidParam5);
-
-    let result = CommandExecutor::instance()
-        .send(Command::Signus(SignusCommand::SetKeyMetadata(
-            wallet_handle,
-            verkey,
-            metadata,
-            Box::new(move |result| {
-                let err = result_to_err_code!(result);
-                cb(command_handle, err)
-            })
-        )));
-
-    result_to_err_code!(result)
-}
-
-/// Retrieves the meta information for the giving key in the wallet.
-///
-/// #Params
-/// command_handle: Command handle to map callback to caller context.
-/// wallet_handle: Wallet handle (created by open_wallet).
-/// verkey - The key (verkey, key id) to retrieve metadata.
-/// cb: Callback that takes command result as parameter.
-///
-/// #Returns
-/// Error Code
-/// cb:
-/// - xcommand_handle: Command handle to map callback to caller context.
-/// - err: Error code.
-/// - metadata - The meta information stored with the key; Can be null if no metadata was saved for this key.
-///
-/// #Errors
-/// Common*
-/// Wallet*
-/// Crypto*
-#[no_mangle]
-pub  extern fn indy_get_key_metadata(command_handle: i32,
-                                     wallet_handle: i32,
-                                     verkey: *const c_char,
-                                     cb: Option<extern fn(xcommand_handle: i32,
-                                                          err: ErrorCode,
-                                                          metadata: *const c_char)>) -> ErrorCode {
-    check_useful_c_str!(verkey, ErrorCode::CommonInvalidParam3);
-    check_useful_c_callback!(cb, ErrorCode::CommonInvalidParam4);
-
-    let result = CommandExecutor::instance()
-        .send(Command::Signus(SignusCommand::GetKeyMetadata(
-            wallet_handle,
-            verkey,
-            Box::new(move |result| {
-                let (err, metadata) = result_to_err_code_1!(result, String::new());
-                let metadata = CStringUtils::string_to_cstring(metadata);
-                cb(command_handle, err, metadata.as_ptr())
-            })
-        )));
-
-    result_to_err_code!(result)
-}
-
 /// Returns ver key (key id) for the given DID.
 /// Note that indy_create_and_store_my_did makes similar wallet record as indy_create_key.
 /// As result we can use returned ver key in all generic crypto and messaging functions.
@@ -522,333 +389,6 @@ pub extern fn indy_get_did_metadata(command_handle: i32,
                 let (err, metadata) = result_to_err_code_1!(result, String::new());
                 let metadata = CStringUtils::string_to_cstring(metadata);
                 cb(command_handle, err, metadata.as_ptr())
-            })
-        )));
-
-    result_to_err_code!(result)
-}
-
-/// Signs a message with a key.
-///
-/// Note to use DID keys with this function you can call indy_key_for_did to get key id (verkey)
-/// for specific DID.
-///
-/// #Params
-/// command_handle: command handle to map callback to user context.
-/// wallet_handle: wallet handler (created by open_wallet).
-/// my_vk: id (verkey) of my key. The key must be created by calling indy_create_key or indy_create_and_store_my_did
-/// message_raw: a pointer to first byte of message to be signed
-/// message_len: a message length
-/// cb: Callback that takes command result as parameter.
-///
-/// #Returns
-/// a signature string
-///
-/// #Errors
-/// Common*
-/// Wallet*
-/// Crypto*
-#[no_mangle]
-pub  extern fn indy_crypto_sign(command_handle: i32,
-                                wallet_handle: i32,
-                                my_vk: *const c_char,
-                                message_raw: *const u8,
-                                message_len: u32,
-                                cb: Option<extern fn(xcommand_handle: i32, err: ErrorCode,
-                                                     signature_raw: *const u8, signature_len: u32)>) -> ErrorCode {
-    check_useful_c_str!(my_vk, ErrorCode::CommonInvalidParam3);
-    check_useful_c_byte_array!(message_raw, message_len, ErrorCode::CommonInvalidParam4, ErrorCode::CommonInvalidParam5);
-    check_useful_c_callback!(cb, ErrorCode::CommonInvalidParam6);
-
-    let result = CommandExecutor::instance()
-        .send(Command::Signus(SignusCommand::CryptoSign(
-            wallet_handle,
-            my_vk,
-            message_raw,
-            Box::new(move |result| {
-                let (err, signature) = result_to_err_code_1!(result, Vec::new());
-                let (signature_raw, signature_len) = vec_to_pointer(&signature);
-                cb(command_handle, err, signature_raw, signature_len)
-            })
-        )));
-
-    result_to_err_code!(result)
-}
-
-/// Verify a signature with a verkey.
-///
-/// Note to use DID keys with this function you can call indy_key_for_did to get key id (verkey)
-/// for specific DID.
-///
-/// #Params
-/// command_handle: command handle to map callback to user context.
-/// their_vk: verkey to use
-/// message_raw: a pointer to first byte of message to be signed
-/// message_len: a message length
-/// signature_raw: a a pointer to first byte of signature to be verified
-/// signature_len: a signature length
-/// cb: Callback that takes command result as parameter.
-///
-/// #Returns
-/// valid: true - if signature is valid, false - otherwise
-///
-/// #Errors
-/// Common*
-/// Wallet*
-/// Ledger*
-/// Crypto*
-#[no_mangle]
-pub  extern fn indy_crypto_verify(command_handle: i32,
-                                  their_vk: *const c_char,
-                                  message_raw: *const u8,
-                                  message_len: u32,
-                                  signature_raw: *const u8,
-                                  signature_len: u32,
-                                  cb: Option<extern fn(xcommand_handle: i32, err: ErrorCode,
-                                                       valid: bool)>) -> ErrorCode {
-    check_useful_c_str!(their_vk, ErrorCode::CommonInvalidParam2);
-    check_useful_c_byte_array!(message_raw, message_len, ErrorCode::CommonInvalidParam3, ErrorCode::CommonInvalidParam4);
-    check_useful_c_byte_array!(signature_raw, signature_len, ErrorCode::CommonInvalidParam5, ErrorCode::CommonInvalidParam6);
-    check_useful_c_callback!(cb, ErrorCode::CommonInvalidParam7);
-
-    let result = CommandExecutor::instance()
-        .send(Command::Signus(SignusCommand::CryptoVerify(
-            their_vk,
-            message_raw,
-            signature_raw,
-            Box::new(move |result| {
-                let (err, valid) = result_to_err_code_1!(result, false);
-                cb(command_handle, err, valid)
-            })
-        )));
-
-    result_to_err_code!(result)
-}
-
-/// Encrypt a message by authenticated-encryption scheme.
-///
-/// Sender can encrypt a confidential message specifically for Recipient, using Sender's public key.
-/// Using Recipient's public key, Sender can compute a shared secret key.
-/// Using Sender's public key and his secret key, Recipient can compute the exact same shared secret key.
-/// That shared secret key can be used to verify that the encrypted message was not tampered with,
-/// before eventually decrypting it.
-///
-/// Recipient only needs Sender's public key, the nonce and the ciphertext to peform decryption.
-/// The nonce doesn't have to be confidential.
-///
-/// Note to use DID keys with this function you can call indy_key_for_did to get key id (verkey)
-/// for specific DID.
-///
-/// #Params
-/// command_handle: command handle to map callback to user context.
-/// wallet_handle: wallet handle (created by open_wallet).
-/// pool_handle: pool handle.
-/// my_vk: id (verkey) of my key. The key must be created by calling indy_create_key or indy_create_and_store_my_did
-/// their_vk: id (verkey) of their key
-/// message_raw: a pointer to first byte of message that to be encrypted
-/// message_len: a message length
-/// cb: Callback that takes command result as parameter.
-///
-/// #Returns
-/// an encrypted message and nonce
-///
-/// #Errors
-/// Common*
-/// Wallet*
-/// Ledger*
-/// Crypto*
-#[no_mangle]
-pub  extern fn indy_crypto_box(command_handle: i32,
-                               wallet_handle: i32,
-                               my_vk: *const c_char,
-                               their_vk: *const c_char,
-                               message_raw: *const u8,
-                               message_len: u32,
-                               cb: Option<extern fn(xcommand_handle: i32, err: ErrorCode,
-                                                    encrypted_msg_raw: *const u8, encrypted_msg_len: u32,
-                                                    nonce_raw: *const u8, nonce_len: u32)>) -> ErrorCode {
-    check_useful_c_str!(my_vk, ErrorCode::CommonInvalidParam3);
-    check_useful_c_str!(their_vk, ErrorCode::CommonInvalidParam4);
-    check_useful_c_byte_array!(message_raw, message_len, ErrorCode::CommonInvalidParam5, ErrorCode::CommonInvalidParam6);
-    check_useful_c_callback!(cb, ErrorCode::CommonInvalidParam7);
-
-    let result = CommandExecutor::instance()
-        .send(Command::Signus(SignusCommand::CryptoBox(
-            wallet_handle,
-            my_vk,
-            their_vk,
-            message_raw,
-            Box::new(move |result| {
-                let (err, encrypted_msg, nonce) = result_to_err_code_2!(result, Vec::new(), Vec::new());
-                let (encrypted_msg_raw, encrypted_msg_len) = vec_to_pointer(&encrypted_msg);
-                let (nonce_raw, nonce_len) = vec_to_pointer(&nonce);
-                cb(command_handle, err, encrypted_msg_raw, encrypted_msg_len, nonce_raw, nonce_len)
-            })
-        )));
-
-    result_to_err_code!(result)
-}
-
-/// Decrypt a message by authenticated-encryption scheme.
-///
-/// Sender can encrypt a confidential message specifically for Recipient, using Sender's public key.
-/// Using Recipient's public key, Sender can compute a shared secret key.
-/// Using Sender's public key and his secret key, Recipient can compute the exact same shared secret key.
-/// That shared secret key can be used to verify that the encrypted message was not tampered with,
-/// before eventually decrypting it.
-///
-/// Recipient only needs Sender's public key, the nonce and the ciphertext to peform decryption.
-/// The nonce doesn't have to be confidential.
-///
-/// Note to use DID keys with this function you can call indy_key_for_did to get key id (verkey)
-/// for specific DID.
-///
-/// #Params
-/// command_handle: command handle to map callback to user context.
-/// wallet_handle: wallet handler (created by open_wallet).
-/// my_vk: id (verkey) of my key. The key must be created by calling indy_create_key or indy_create_and_store_my_did
-/// their_vk: id (verkey) of their key
-/// encrypted_msg_raw: a pointer to first byte of message that to be decrypted
-/// encrypted_msg_len: a message length
-/// nonce_raw: a pointer to first byte of nonce that encrypted message
-/// nonce_len: a nonce length
-/// cb: Callback that takes command result as parameter.
-///
-/// #Returns
-/// decrypted message
-///
-/// #Errors
-/// Common*
-/// Wallet*
-/// Crypto*
-#[no_mangle]
-pub  extern fn indy_crypto_box_open(command_handle: i32,
-                                    wallet_handle: i32,
-                                    my_vk: *const c_char,
-                                    their_vk: *const c_char,
-                                    encrypted_msg_raw: *const u8,
-                                    encrypted_msg_len: u32,
-                                    nonce_raw: *const u8,
-                                    nonce_len: u32,
-                                    cb: Option<extern fn(xcommand_handle: i32, err: ErrorCode,
-                                                         decrypted_msg_raw: *const u8, decrypted_msg_len: u32)>) -> ErrorCode {
-    check_useful_c_str!(my_vk, ErrorCode::CommonInvalidParam3);
-    check_useful_c_str!(their_vk, ErrorCode::CommonInvalidParam4);
-    check_useful_c_byte_array!(encrypted_msg_raw, encrypted_msg_len, ErrorCode::CommonInvalidParam5, ErrorCode::CommonInvalidParam6);
-    check_useful_c_byte_array!(nonce_raw, nonce_len, ErrorCode::CommonInvalidParam7, ErrorCode::CommonInvalidParam8);
-    check_useful_c_callback!(cb, ErrorCode::CommonInvalidParam9);
-
-    let result = CommandExecutor::instance()
-        .send(Command::Signus(SignusCommand::CryptoBoxOpen(
-            wallet_handle,
-            my_vk,
-            their_vk,
-            encrypted_msg_raw,
-            nonce_raw,
-            Box::new(move |result| {
-                let (err, decrypted_msg) = result_to_err_code_1!(result, Vec::new());
-                let (decrypted_msg_raw, decrypted_msg_len) = vec_to_pointer(&decrypted_msg);
-                cb(command_handle, err, decrypted_msg_raw, decrypted_msg_len)
-            })
-        )));
-
-    result_to_err_code!(result)
-}
-
-/// Encrypts a message by anonymous-encryption scheme.
-///
-/// Sealed boxes are designed to anonymously send messages to a Recipient given its public key.
-/// Only the Recipient can decrypt these messages, using its private key.
-/// While the Recipient can verify the integrity of the message, it cannot verify the identity of the Sender.
-///
-/// Note to use DID keys with this function you can call indy_key_for_did to get key id (verkey)
-/// for specific DID.
-///
-/// #Params
-/// command_handle: command handle to map callback to user context.
-/// their_vk: id (verkey) of their key
-/// message_raw: a pointer to first byte of message that to be encrypted
-/// message_len: a message length
-/// cb: Callback that takes command result as parameter.
-///
-/// #Returns
-/// an encrypted message
-///
-/// #Errors
-/// Common*
-/// Wallet*
-/// Ledger*
-/// Crypto*
-#[no_mangle]
-pub  extern fn indy_crypto_box_seal(command_handle: i32,
-                                    their_vk: *const c_char,
-                                    message_raw: *const u8,
-                                    message_len: u32,
-                                    cb: Option<extern fn(xcommand_handle: i32, err: ErrorCode,
-                                                         encrypted_msg_raw: *const u8, encrypted_msg_len: u32)>) -> ErrorCode {
-    check_useful_c_str!(their_vk, ErrorCode::CommonInvalidParam2);
-    check_useful_c_byte_array!(message_raw, message_len, ErrorCode::CommonInvalidParam3, ErrorCode::CommonInvalidParam4);
-    check_useful_c_callback!(cb, ErrorCode::CommonInvalidParam5);
-
-    let result = CommandExecutor::instance()
-        .send(Command::Signus(SignusCommand::CryptoBoxSeal(
-            their_vk,
-            message_raw,
-            Box::new(move |result| {
-                let (err, encrypted_msg) = result_to_err_code_1!(result, Vec::new());
-                let (encrypted_msg_raw, encrypted_msg_len) = vec_to_pointer(&encrypted_msg);
-                cb(command_handle, err, encrypted_msg_raw, encrypted_msg_len)
-            })
-        )));
-
-    result_to_err_code!(result)
-}
-
-/// Decrypts a message by anonymous-encryption scheme.
-///
-/// Sealed boxes are designed to anonymously send messages to a Recipient given its public key.
-/// Only the Recipient can decrypt these messages, using its private key.
-/// While the Recipient can verify the integrity of the message, it cannot verify the identity of the Sender.
-///
-/// Note to use DID keys with this function you can call indy_key_for_did to get key id (verkey)
-/// for specific DID.
-///
-/// #Params
-/// command_handle: command handle to map callback to user context.
-/// wallet_handle: wallet handler (created by open_wallet).
-/// my_vk: id (verkey) of my key. The key must be created by calling indy_create_key or indy_create_and_store_my_did
-/// encrypted_msg_raw: a pointer to first byte of message that to be decrypted
-/// encrypted_msg_len: a message length
-/// cb: Callback that takes command result as parameter.
-///
-/// #Returns
-/// decrypted message
-///
-/// #Errors
-/// Common*
-/// Wallet*
-/// Crypto*
-#[no_mangle]
-pub  extern fn indy_crypto_box_seal_open(command_handle: i32,
-                                         wallet_handle: i32,
-                                         my_vk: *const c_char,
-                                         encrypted_msg_raw: *const u8,
-                                         encrypted_msg_len: u32,
-                                         cb: Option<extern fn(xcommand_handle: i32, err: ErrorCode,
-                                                              decrypted_msg_raw: *const u8, decrypted_msg_len: u32)>) -> ErrorCode {
-    check_useful_c_str!(my_vk, ErrorCode::CommonInvalidParam3);
-    check_useful_c_byte_array!(encrypted_msg_raw, encrypted_msg_len, ErrorCode::CommonInvalidParam4, ErrorCode::CommonInvalidParam5);
-    check_useful_c_callback!(cb, ErrorCode::CommonInvalidParam6);
-
-    let result = CommandExecutor::instance()
-        .send(Command::Signus(SignusCommand::CryptoBoxSealOpen(
-            wallet_handle,
-            my_vk,
-            encrypted_msg_raw,
-            Box::new(move |result| {
-                let (err, decrypted_msg) = result_to_err_code_1!(result, Vec::new());
-                let (decrypted_msg_raw, decrypted_msg_len) = vec_to_pointer(&decrypted_msg);
-                cb(command_handle, err, decrypted_msg_raw, decrypted_msg_len)
             })
         )));
