+++ libindy/src/commands/mod.rs
@@ -4,6 +4,7 @@ mod utils;
 pub mod agent;
 #[allow(unused_variables)] /* FIXME */
 pub mod anoncreds;
+pub mod crypto;
 pub mod ledger;
 pub mod pool;
 pub mod signus;
@@ -12,6 +13,7 @@ pub mod pairwise;
 
 use commands::agent::{AgentCommand, AgentCommandExecutor};
 use commands::anoncreds::{AnoncredsCommand, AnoncredsCommandExecutor};
+use commands::crypto::{CryptoCommand, CryptoCommandExecutor};
 use commands::ledger::{LedgerCommand, LedgerCommandExecutor};
 use commands::pool::{PoolCommand, PoolCommandExecutor};
 use commands::signus::{SignusCommand, SignusCommandExecutor};
@@ -36,6 +38,7 @@ pub enum Command {
     Exit,
     Agent(AgentCommand),
     Anoncreds(AnoncredsCommand),
+    Crypto(CryptoCommand),
     Ledger(LedgerCommand),
     Pool(PoolCommand),
     Signus(SignusCommand),
@@ -75,6 +78,7 @@ impl CommandExecutor {
 
                 let agent_command_executor = AgentCommandExecutor::new(signus_service.clone(), wallet_service.clone());
                 let anoncreds_command_executor = AnoncredsCommandExecutor::new(anoncreds_service.clone(), pool_service.clone(), wallet_service.clone());
+                let crypto_command_executor = CryptoCommandExecutor::new(wallet_service.clone(), signus_service.clone());
                 let ledger_command_executor = LedgerCommandExecutor::new(pool_service.clone(), signus_service.clone(), wallet_service.clone(), ledger_service.clone());
                 let pool_command_executor = PoolCommandExecutor::new(pool_service.clone());
                 let signus_command_executor = SignusCommandExecutor::new(pool_service.clone(), wallet_service.clone(), signus_service.clone(), ledger_service.clone());
@@ -84,39 +88,43 @@ impl CommandExecutor {
                 loop {
                     match receiver.recv() {
                         Ok(Command::Agent(cmd)) => {
-                            info!(target: "command_executor", "AgentCommand command received");
+                            info!("AgentCommand command received");
                             agent_command_executor.execute(cmd);
                         }
                         Ok(Command::Anoncreds(cmd)) => {
-                            info!(target: "command_executor", "AnoncredsCommand command received");
+                            info!("AnoncredsCommand command received");
                             anoncreds_command_executor.execute(cmd);
                         }
+                        Ok(Command::Crypto(cmd)) => {
+                            info!("CryptoCommand command received");
+                            crypto_command_executor.execute(cmd);
+                        }
                         Ok(Command::Ledger(cmd)) => {
-                            info!(target: "command_executor", "LedgerCommand command received");
+                            info!("LedgerCommand command received");
                             ledger_command_executor.execute(cmd);
                         }
                         Ok(Command::Pool(cmd)) => {
-                            info!(target: "command_executor", "PoolCommand command received");
+                            info!("PoolCommand command received");
                             pool_command_executor.execute(cmd);
                         }
                         Ok(Command::Signus(cmd)) => {
-                            info!(target: "command_executor", "SignusCommand command received");
+                            info!("SignusCommand command received");
                             signus_command_executor.execute(cmd);
                         }
                         Ok(Command::Wallet(cmd)) => {
-                            info!(target: "command_executor", "WalletCommand command received");
+                            info!("WalletCommand command received");
                             wallet_command_executor.execute(cmd);
                         }
                         Ok(Command::Pairwise(cmd)) => {
-                            info!(target: "command_executor", "PairwiseCommand command received");
+                            info!("PairwiseCommand command received");
                             pairwise_command_executor.execute(cmd);
                         }
                         Ok(Command::Exit) => {
-                            info!(target: "command_executor", "Exit command received");
+                            info!("Exit command received");
                             break
                         }
                         Err(err) => {
-                            error!(target: "command_executor", "Failed to get command!");
+                            error!("Failed to get command!");
                             panic!("Failed to get command! {:?}", err)
                         }
                     }
