+++ libindy/tests/signus.rs
@@ -37,201 +37,6 @@ fn indy_set_did_metadata_works() {
 mod high_cases {
     use super::*;
 
-    mod create_key {
-        use super::*;
-        use rust_base58::FromBase58;
-
-        #[test]
-        fn indy_create_key_works_for_seed() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            let verkey = SignusUtils::create_key(wallet_handle, Some(MY1_SEED)).unwrap();
-            assert_eq!(verkey.from_base58().unwrap().len(), 32);
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-
-        #[test]
-        fn indy_create_key_works_without_seed() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            let verkey = SignusUtils::create_key(wallet_handle, None).unwrap();
-            assert_eq!(verkey.from_base58().unwrap().len(), 32);
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-
-        #[test]
-        fn indy_create_key_works_for_invalid_wallet_handle() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            let invalid_wallet_handle = wallet_handle + 1;
-            let res = SignusUtils::create_key(invalid_wallet_handle, None);
-            assert_eq!(ErrorCode::WalletInvalidHandle, res.unwrap_err());
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-    }
-
-    mod set_key_metadata {
-        use super::*;
-
-        #[test]
-        fn indy_set_key_metadata_works() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            SignusUtils::set_key_metadata(wallet_handle, VERKEY, METADATA).unwrap();
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-
-        #[test]
-        fn indy_set_key_metadata_works_for_replace() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            SignusUtils::set_key_metadata(wallet_handle, VERKEY, METADATA).unwrap();
-            let metadata = SignusUtils::get_key_metadata(wallet_handle, VERKEY).unwrap();
-            assert_eq!(METADATA.to_string(), metadata);
-
-            let new_metadata = "updated metadata";
-            SignusUtils::set_key_metadata(wallet_handle, VERKEY, new_metadata).unwrap();
-            let updated_metadata = SignusUtils::get_key_metadata(wallet_handle, VERKEY).unwrap();
-            assert_eq!(new_metadata, updated_metadata);
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-
-        #[test]
-        fn indy_set_key_metadata_works_for_invalid_handle() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            let invalid_wallet_handle = wallet_handle + 1;
-            let res = SignusUtils::set_key_metadata(invalid_wallet_handle, VERKEY, METADATA);
-            assert_eq!(ErrorCode::WalletInvalidHandle, res.unwrap_err());
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-
-        #[test]
-        fn indy_set_key_metadata_works_for_empty_string() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            SignusUtils::set_key_metadata(wallet_handle, VERKEY, "").unwrap();
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-
-
-        #[test]
-        fn indy_set_key_metadata_works_for_invalid_key() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            let res = SignusUtils::set_key_metadata(wallet_handle, INVALID_BASE58_VERKEY, METADATA);
-            assert_eq!(ErrorCode::CommonInvalidStructure, res.unwrap_err());
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-    }
-
-    mod get_key_metadata {
-        use super::*;
-
-        #[test]
-        fn indy_get_key_metadata_works() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            SignusUtils::set_key_metadata(wallet_handle, VERKEY, METADATA).unwrap();
-
-            let metadata = SignusUtils::get_key_metadata(wallet_handle, VERKEY).unwrap();
-            assert_eq!(METADATA.to_string(), metadata);
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-
-        #[test]
-        fn indy_get_key_metadata_works_for_empty_string() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            SignusUtils::set_key_metadata(wallet_handle, VERKEY, "").unwrap();
-
-            let metadata = SignusUtils::get_key_metadata(wallet_handle, VERKEY).unwrap();
-            assert_eq!("", metadata);
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-
-        #[test]
-        fn indy_get_key_metadata_works_for_no_metadata() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            let res = SignusUtils::get_key_metadata(wallet_handle, VERKEY);
-            assert_eq!(ErrorCode::WalletNotFoundError, res.unwrap_err());
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-
-        #[test]
-        fn indy_get_key_metadata_works_for_invalid_handle() {
-            TestUtils::cleanup_storage();
-
-            let wallet_handle = WalletUtils::create_and_open_wallet(POOL, None).unwrap();
-
-            SignusUtils::set_key_metadata(wallet_handle, VERKEY, METADATA).unwrap();
-
-            let invalid_invalid_handle = wallet_handle + 1;
-            let res = SignusUtils::get_key_metadata(invalid_invalid_handle, VERKEY);
-            assert_eq!(ErrorCode::WalletInvalidHandle, res.unwrap_err());
-
-            WalletUtils::close_wallet(wallet_handle).unwrap();
-
-            TestUtils::cleanup_storage();
-        }
-    }
-
     mod key_for_did {
         use super::*;
